requirements.txt
anyjson==0.3.3
circus==0.14.0
circus-web==1.0.0
decorator==4.0.10
discogs-client==2.2.1
Django==1.10
django-debug-toolbar==1.5
django-debugtools==1.7
django-filter==0.14.0
django-grappelli==2.8.2
django-redis==4.4.4
django-redis-sentinel==1.0
djangorestframework==3.4.6
gevent==1.1.2
greenlet==0.4.10
httpie==0.9.6
iowait==0.2
ipython==5.1.0
ipython-genutils==0.1.0
Mako==1.0.4
Markdown==2.6.6
MarkupSafe==0.23
musicbrainzngs==0.6
oauthlib==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
prompt-toolkit==1.0.7
psutil==4.3.1
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
pytz==2016.6.1
pyzmq==15.4.0
redis==2.10.5
requests==2.11.1
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
sqlparse==0.2.1
tomako==0.1.0.post1
TornadIO2==0.0.3
tornado==4.4.1
traitlets==4.2.2
wcwidth==0.1.7
